hopefully fix build on CI

+++ onnx/defs/tensor/defs.cc
@@ -778,7 +778,7 @@ ONNX_OPERATOR_SET_SCHEMA(
             }
           } else { // no value available for 'split'
             if (num_outputs_attr) {
-              const int num_outputs = num_outputs_attr->i();
+              const long int num_outputs = num_outputs_attr->i();
               if (num_outputs < 1) {
                 fail_shape_inference("Attribute `num_outputs` value cannot be lower than 1");
               }
